requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
pytest